LPS-76837 Move tests to correct packages
